oeps

+++ src/com/capgemini/se_track/model/Product.java
@@ -1,21 +1,19 @@
 package com.capgemini.se_track.model;
 
-import java.math.BigDecimal;
-
 public abstract class Product {
     private ProductName name;
-    private BigDecimal price;
+    private double price;
 
     public Product(ProductName name, double price) {
         this.name = name;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
     }
 
     public String getName() {
         return name.toString();
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
